Scale down our largest two font sizes on mobile.
Bump $font__size-xxl and $font__size-xl down one step on the type scale when we're on smaller screens. This helps preserve a reasonable character count when we're dealing with narrower viewports.

This _does_ mean that on mobile, h2s and h3s are the same type size, but it doesn't bother me because h2s will still have the short divider line above them to signify a greater hierarchy.