Consolidate editor styles
A followup to #22.

This PR removes our secondary `style-editor-frame.css` file, and consolidates all of our editor styles into `style-editor.css`. Initially, the standard editor stylesheet was limited in scope to individual blocks, which meant we were unable to style the general alignment of the page, as well as items like the page title and default appender.

The scope of `style-editor.css` is being expanded in https://github.com/WordPress/gutenberg/pull/10956 , so we're now able to style those elements in a single stylesheet.

Some minor style tweaks were necessary to make our original styles work in this context, but for the most part this is identical to #22.