Add default widget styles, clean up widget block styles.
Add styles so that the default widgets appear similar to the rest of the theme.

Also, revise the Latest Posts, Archives, and Categories block styles slightly to sync them up with the widgets. This commit also removes the commas and periods that were added at the end of list items in these blocks, as those didn't make sense in many other languages.