Fix additional issues with columns block.
The previous fix to nested blocks in general did not sufficiently address the cascading of properties. Recent refactors to the columns block in Gutenberg upstream made this visible.

This PR fixes those issues by increasing the specificity of a few of the rules that would also target any nested blocks.