Remove hyphens rule for cover image text
On the front end, we add hyphens to  broken lines via `hyphens: auto`. This was intended to provide more natural line breaks, but in effect I've found that it makes the browser a little over-eager to add hyphens. Removing this rule results in more expected behavior, and better matches what you'd see in the eidtor (where there are no hyphens).