Add a notice about the transition to Trac
Issues and PRs in this repository are currently in the process of being transitioned to Trac. To make sure that's clear to people, it'd be helpful to include notices here to point contributors over there.

This commit includes a note in the main `README.md`, as well as issue & pull request templates, so that people can be notified of the transition before they submit an issue.